+++ routes/TabNavigation.tsx
@@ -7,7 +7,7 @@ import { Button, Pressable } from 'react-native';
 
 import Home from "@/screens/Home";
 import Transaction from "@/screens/transaction/Transaction";
-import Statistic from "@/screens/Statistic";
+import Statistic from "@/screens/statistic/Statistic";
 import Profile from "@/screens/Profile";
 import { getAuth, signOut } from "firebase/auth";
 import type { NativeStackNavigationProp } from "@react-navigation/native-stack";
