+++ routes/authStack.tsx
@@ -2,15 +2,15 @@ import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { TransactionProvider } from "@/screens/transaction/TransactionContext";
-import Register from "@/screens/Register";
+import Register from "@/screens/authentication/Register";
 import Home from "@/screens/Home";
-import Login from "@/screens/Login";
+import Login from "@/screens/authentication/Login";
 import TabNavigation from "@/routes/TabNavigation";
-import LoginPhoneNumber from "@/screens/LoginPhoneNumber";
-import ForgetPassword from "@/screens/ForgetPassword";
+import LoginPhoneNumber from "@/screens/authentication/LoginPhoneNumber";
+import ForgetPassword from "@/screens/authentication/ForgetPassword";
 import DrawerNavigation from "@/routes/DrawerNavigation";
-import IncomeStatistic from "@/screens/IncomeStatistic";
-import LoanStatistic from "@/screens/LoanStatistic";
+import IncomeStatistic from "@/screens/statistic/IncomeStatistic";
+import LoanStatistic from "@/screens/statistic/LoanStatistic";
 import AddTransaction from "@/screens/transaction/AddTransaction";
 import Transaction from "@/screens/transaction/Transaction";
 import TransactionDetail from "@/screens/transaction/TransactionDetail";
