+++ screens/authentication/ForgetPassword.tsx
@@ -18,7 +18,6 @@ import {
 import type { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { useNavigation } from "@react-navigation/native";
 import { getAuth, signInWithEmailAndPassword, sendEmailVerification, sendPasswordResetEmail } from 'firebase/auth';
-import app from '../firebaseConfig';
 
 export default function ForgetPassword() {
     const [email, setEmail] = useState<string | undefined>();
@@ -41,7 +40,7 @@ export default function ForgetPassword() {
     return (
         <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
             <SafeAreaView style={styles.overall}>
-                <Image source={require('../assets/images/logo1.png')} style={styles.logo} />
+                <Image source={require('@/assets/images/logo1.png')} style={styles.logo} />
                 <View style={styles.container}>
                 
                     <Text style={styles.title}>Forget Password</Text>
