+++ screens/authentication/LoginPhoneNumber.tsx
@@ -17,12 +17,13 @@ import {
 
 import type { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { useNavigation } from "@react-navigation/native";
-import { getAuth, signInWithEmailAndPassword, sendEmailVerification, sendPasswordResetEmail } from 'firebase/auth';
-import app from '../firebaseConfig';
-import { getDatabase , ref, get} from "firebase/database";
+import { getAuth, signInWithEmailAndPassword, sendEmailVerification } from 'firebase/auth';
+import app from '../../firebaseConfig';
+import PhoneInput from 'react-native-phone-input'; 
 
-export default function Login() {
+export default function LoginPhoneNumber() {
     const [email, setEmail] = useState<string | undefined>();
+    const [phoneNumber, setPhoneNumber] = useState<string | undefined>();
     const [password, setPassword] = useState<string | undefined>();
     const [loading, setLoading] = useState<boolean>(false);
     const [showPassword, setShowPassword] = useState<boolean>(false);
@@ -30,36 +31,40 @@ export default function Login() {
     const navigation = useNavigation<NativeStackNavigationProp<any>>();
 
     const handleLogin = async () => {
-        if (email && password) {
-            setLoading(true);
-            try {
-                const auth = getAuth(app);
-                const response = await signInWithEmailAndPassword(auth, email, password);
+        // if (phoneNumber) {
+        //     setLoading(true);
+        //     try {
 
-                if (response.user && response.user.emailVerified) {
-                    
-                    setLoading(false);
-                    navigation.replace('DrawerNavigation');
-                } else {
-                    setModalVisibility(true);
+        //     }
+        // }
+        // if (email && password) {
+        //     setLoading(true);
+        //     try {
+        //         const auth = getAuth(app);
+        //         const response = await signInWithEmailAndPassword(auth, email, password);
+
+        //         if (response.user && response.user.emailVerified) {
+        //             setLoading(false);
+        //             navigation.replace('TabNavigation');
+        //         } else {
+        //             setModalVisibility(true);
         
-                }
-            } catch (e) {
-                setLoading(false);
-                let err = "Try Again";
-                if (e instanceof Error && e.message) {
-                    err = e.message;
-                }
-                Alert.alert("Error", "Invalid Email or Password");
-            }
-        }
+        //         }
+        //     } catch (e) {
+        //         setLoading(false);
+        //         let err = "Try Again";
+        //         if (e instanceof Error && e.message) {
+        //             err = e.message;
+        //         }
+        //         Alert.alert("Error", "Invalid Email or Password");
+        //     }
+        // }
     };
 
     const resendVerificationEmail = () => {
         const auth = getAuth(app);
         const user = auth.currentUser;
         if (user) {
-          console.log(1)
           sendEmailVerification(user)
           .then(() => Alert.alert("Verification Email Resent", "Please check your email"))
           .catch((e) => {Alert.alert("Error sending email", "Please try again later")});
@@ -76,12 +81,12 @@ export default function Login() {
               setModalVisibility(false);
               navigation.reset({
                 index: 0,
-                routes: [{name: 'DrawerNavigation'}],
+                routes: [{name: 'TabNavigation'}],
               })
             }
           })
         }
-      }
+    }
 
     useEffect(() => {
         if (modalVisibility) {
@@ -93,9 +98,10 @@ export default function Login() {
     }, [modalVisibility]);
 
     return (
+        
         <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
             <SafeAreaView style={styles.overall}>
-                <Image source={require('../assets/images/logo1.png')} style={styles.logo} />
+                <Image source={require('@/assets/images/logo1.png')} style={styles.logo} />
                 <View style={styles.container}>
                 <Modal
                 animationType="slide"
@@ -113,33 +119,12 @@ export default function Login() {
                 </Modal>
                     <Text style={styles.title}>Login</Text>
                     <View style={styles.content}>
-                        <Text style={styles.contentText}>Email</Text>
-                        <TextInput
-                            style={styles.input}
-                            placeholder="Email"
-                            placeholderTextColor="#aaa"
-                            value={email}
-                            onChangeText={setEmail}
-                            inputMode="email"
-                            autoCapitalize="none"
-                        />
-                        <Text style={styles.contentText}>Password</Text>
-                        <View style={styles.passwordContainer}>
-                            <TextInput
-                                style={styles.passwordInput}
-                                placeholder="Password"
-                                placeholderTextColor="#aaa"
-                                value={password}
-                                onChangeText={setPassword}
-                                secureTextEntry={!showPassword}
-                            />
-                            <TouchableOpacity
-                                onPress={() => setShowPassword(!showPassword)}
-                                style={styles.showPasswordButton}
-                            >
-                                <Text style={styles.showPasswordText}>{showPassword ? "Hide" : "Show"}</Text>
-                            </TouchableOpacity>
-                        </View>
+                        <Text style={styles.contentText}>PhoneNumber</Text>
+                        <PhoneInput 
+                  initialValue={phoneNumber}
+                  initialCountry="sg"
+                  onChangePhoneNumber={(number) => setPhoneNumber(number)}
+                  style = {styles.input}/>
                         {loading ? (
                             <ActivityIndicator size="large" color="#3498db" />
                         ) : (
@@ -147,13 +132,6 @@ export default function Login() {
                                 <Text style={styles.loginButtonText}>Login</Text>
                             </TouchableOpacity>
                         )}
-                        
-                        <View style = {styles.forgetPasswordContainer}>
-                            <TouchableOpacity onPress={() => navigation.navigate('ForgetPassword')}>
-                            <Text style={styles.forgetPasswordNavigation}>Forget Password?</Text>
-                            </TouchableOpacity>
-                        </View>
-                        
                         <View style={styles.registerContainer}>
                             <Text style={styles.navigateTitle}>No account?</Text>
                             <TouchableOpacity onPress={() => navigation.navigate('Register')}>
@@ -162,7 +140,7 @@ export default function Login() {
                         </View>
                     </View>
                 </View>
-                <Button title="Login Using Phone Number" onPress={() => navigation.navigate('LoginPhoneNumber')}/>
+                <Button title="Login Using Email" onPress={() => navigation.navigate('Login')}/>
             </SafeAreaView>
         </TouchableWithoutFeedback>
     );
@@ -261,7 +239,7 @@ const styles = StyleSheet.create({
         flexDirection: 'row',
         justifyContent: 'center',
         alignItems: 'center',
-        marginTop: 10,
+        marginTop: 20,
     },
     navigateTitle: {
         fontSize: 14,
@@ -272,15 +250,5 @@ const styles = StyleSheet.create({
         color: '#3498db',
         marginLeft: 5,
         fontWeight: 'bold',
-    },
-    forgetPasswordContainer: {
-        flexDirection:'row',
-        justifyContent:'flex-end',
-        alignItems:'flex-end',
-        marginTop: 15,
-    },
-    forgetPasswordNavigation: {
-        fontSize: 14,
-        color: '#3498db',
     }
 });
